New readings page using normal templates + Tailwind
Rebuild the readings page to use Go templates (non-Ace) like other more
modern pages and Tailwind everywhere.
