+++ views/reading/index.tmpl.html
@@ -1,0 +1,103 @@
+{{- template "layouts/atoms.tmpl.html" . -}}
+
+{{- define "atoms_content" -}}
+
+<div class="mb-12 mt-0 md:mb-24 md:mt-16 px-4">
+    <h1 class="text-slate-800 font-normal font-serif my-8 text-center text-8xl tracking-tighter">
+        Reading
+    </h1>
+
+    <div class="container max-w-[625px] mx-auto
+            prose prose-lg
+            prose-a:border-b-[1px] prose-a:border-slate-500 prose-a:font-sans prose-a:no-underline
+            hover:prose-a:border-slate-200
+            prose-p:text-center prose-p:italic
+            ">
+        <p>
+            Books I've read. This is not Uber — consider four stars very good,
+            and five stars exceptionally so. Synced from Goodreads until their API
+            is finally removed. One paragraph reviews.
+        </p>
+    </div>
+</div>
+
+<div class="container max-w-[1150px] mx-auto mt-8 px-4">
+    <div class="md:flex">
+        <div class="md:basis-[115px] md:flex-grow-0 md:flex-shrink-0">
+            <div class="font-mono my-1 leading-normal text-md tracking-tighter md:sticky md:text-right md:top-5">
+                <div class="hidden md:block md:px-4">
+                    <ul>
+                        {{range .ReadingsByYear}}
+                            <li class="pb-1">
+                                <a href="#year_{{.Year}}">{{.Year}}</a>
+                            </li>
+                        {{- end -}}
+                    </ul>
+                </div>
+            </div>
+        </div>
+        <div class="pb-8 md:border-l-[1px] md:flex-grow md:min-w-0">
+            {{range .ReadingsByYear}}
+                <div class="border-b-[1px] py-6 md:pl-8">
+                    <h2 id="year_{{.Year}}" class="font-bold text-md tracking-tighter">{{.Year}}</h2>
+                </div>
+
+                <ul>
+                    {{- range .Readings -}}
+                        <div class="border-b-[1px] flex flex-col md:pl-8">
+                            <div class="flex gap-6 py-6">
+                                <div class="flex-grow {{if ne .Review ""}} lg:basis-[250px] lg:flex-grow-0 lg:flex-shrink-0 {{end}}">
+                                <!-- <div class="flex-grow"> -->
+                                    <div class="flex flex-col gap-1">
+                                        <div class="font-semibold text-sm">{{.Title}}</div>
+                                        <div class="italic text-xs">{{.AuthorsDisplay}}</div>
+                                    </div>
+                                </div>
+                                <!-- needs at least 80px to support five star reviews -->
+                                <div class="basis-[80px] flex-grow-0 flex-shrink-0 text-xs">
+                                {{if ne .Rating 0}}
+                                    {{ToStars .Rating}}
+                                {{end}}
+                                </div>
+                                <div class="basis-[40px] flex-grow-0 flex-shrink-0 hidden text-xs sm:block">
+                                    {{if eq .NumPages 0}}
+                                        <span class="text-slate-400">—</span>
+                                    {{else}}
+                                        {{.NumPages}}p.
+                                    {{end}}
+                                </div>
+                                <div class="basis-[40px] flex-grow-0 flex-shrink-0 hidden text-xs sm:block">
+                                    <span>{{FormatTime .ReadAt "Jan 2"}}</span>
+                                </div>
+                                {{if .ReviewHTML}}
+                                    <div class="hidden
+                                                hyphens-auto
+                                                mt-[-3px]
+                                                prose
+                                                prose-p:font-serif prose-p:leading-5 prose-p:text-sm
+                                                lg:block
+                                                ">
+                                        {{.ReviewHTML}}
+                                    </div>
+                                {{end}}
+                            </div>
+                            {{if .ReviewHTML}}
+                                <div class="hyphens-auto
+                                            max-w-none
+                                            pb-6
+                                            prose
+                                            prose-p:font-serif prose-p:leading-5 prose-p:text-sm
+                                            lg:hidden
+                                            ">
+                                    {{.ReviewHTML}}
+                                </div>
+                            {{end}}
+                        </div>
+                    {{- end -}}
+                </ul>
+            {{- end -}}
+        </div>
+    </div>
+</div>
+
+{{- end -}}
